tools/game_ocr/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "game-data-ocr"
version = "0.1.0"
description = "Screenshot-first OCR extractor for NHL menu and postgame screens"
readme = "Game OCR Document.md"
requires-python = ">=3.11"
dependencies = [
  "opencv-python",
  "Pillow",
  "pydantic>=2",
  "PyYAML",
  "rapidocr_onnxruntime",
  "typer",
]

[project.scripts]
game-ocr = "game_ocr.cli:app"

[tool.setuptools]
packages = ["game_ocr"]

[tool.setuptools.package-data]
game_ocr = ["configs/roi/*.yaml"]
